bug fix - client deletion

+++ src/main/java/com/cydeo/accountingsimplified/controller/ClientVendorController.java
@@ -4,10 +4,12 @@ import com.cydeo.accountingsimplified.dto.ClientVendorDto;
 import com.cydeo.accountingsimplified.enums.ClientVendorType;
 import com.cydeo.accountingsimplified.service.AddressService;
 import com.cydeo.accountingsimplified.service.ClientVendorService;
+import com.cydeo.accountingsimplified.service.InvoiceService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 import java.util.Arrays;
@@ -17,10 +19,12 @@ import java.util.Arrays;
 public class ClientVendorController {
 
     private final ClientVendorService clientVendorService;
+    private final InvoiceService invoiceService;
     private final AddressService addressService;
 
-    public ClientVendorController(ClientVendorService clientVendorService, AddressService addressService) {
+    public ClientVendorController(ClientVendorService clientVendorService, InvoiceService invoiceService, AddressService addressService) {
         this.clientVendorService = clientVendorService;
+        this.invoiceService = invoiceService;
         this.addressService = addressService;
     }
 
@@ -71,13 +75,17 @@ public class ClientVendorController {
     }
 
     @GetMapping(value = "/delete/{clientVendorId}")
-    public String activateCompany(@PathVariable("clientVendorId") Long clientVendorId) {
+    public String deleteClient(@PathVariable("clientVendorId") Long clientVendorId, RedirectAttributes redirAttrs) {
+        if (invoiceService.checkIfInvoiceExist(clientVendorId)){
+            redirAttrs.addFlashAttribute("error", "Can not be deleted...You have invoices with this client/vendor");
+            return "redirect:/clientVendors/list";
+        }
         clientVendorService.delete(clientVendorId);
         return "redirect:/clientVendors/list";
     }
 
     @ModelAttribute
-    public void getTitle(Model model){
+    public void commonAttributes(Model model){
         model.addAttribute("countries",addressService.getCountryList() );
         model.addAttribute("clientVendorTypes", Arrays.asList(ClientVendorType.values()));
         model.addAttribute("title", "Cydeo Accounting-Client/Vendor");
